Crud Changes

+++ tables/algae.php
@@ -1,9 +1,9 @@
 <?PHP
 include("../header.php");
 include("../db_connection/connection.php");
+include("../constants.php");
 
-
-if(isset($_GET['type'])=='delete'){
+if(isset($_GET['type']) && $_GET['type']=='delete'){
   
   
   $result=mysqli_query($con,"delete from algue where id ={$_GET['id']}");
@@ -17,6 +17,10 @@ if(isset($_GET['type'])=='delete'){
 
 
 <div class="jumbotron">
+<div class="d-flex justify-content-start">
+  <a class="btn btn-outline-success mb-1" href="../Algae.php">
+                <i class="fa fa-plus-circle" aria-hidden="true"></i>
+</a></div>
 <table class="table table-hover table-condensed">
     <thead>
             <th>id</th>
@@ -28,6 +32,7 @@ if(isset($_GET['type'])=='delete'){
             <th>collector</th>
             <th>collection_number</th>
             <th>year</th>
+            <th>Actions</th>
            
 
     <thead>
@@ -47,6 +52,7 @@ if(isset($_GET['type'])=='delete'){
         <td>".$row['collection_number']."</td>
         <td>".$row['year']."</td>
         <td> <a class='btn btn-danger' href='?id={$row['id']}&type=delete'>delete</a></td>
+        <td><a class='btn btn-primary' href='{$ROOT}/Algae.php?id={$row['id']}&type=update'>update</a></td>
         </tr>";
     }
